--- js/functions.js
@@ -1,28 +1,0 @@
-//Функция для проверки длины строки. Она нам пригодится для валидации формы:
-const checkLenght = (string,number) => string.length <= number;
-//Функция для проверки, является ли строка палиндромом:
-const checkPalindrom = (string) => {
-  const tempString = string.toLowerCase().replaceAll(' ', '');
-
-  let reverseString = '';
-  for (let i = tempString.length - 1; i >= 0; i--) {
-    reverseString += tempString.at(i);
-  }
-  return tempString === reverseString;
-};
-//Функция, которая принимает строку, извлекает содержащиеся в ней цифры от 0 до 9 и возвращает их в виде целого положительного числа. Если в строке нет ни одной цифры, функция должна вернуть NaN:
-const extractNumber = (string)=> {
-  if (typeof string === 'number') {
-    return string;
-  }
-  let result = '';
-  for (let i = 0; i < string.length; i++) {
-    if (!Number.isNaN(parseInt(string.at(i), 10))) {
-      result += string.at(i);
-    }
-  }
-  return parseInt(result, 10);
-};
-//Функция, которая принимает три параметра: исходную строку, минимальную длину и строку с добавочными символами — и возвращает исходную строку, дополненную указанными символами до заданной длины. Символы добавляются в начало строки. Если исходная строка превышает заданную длину, она не должна обрезаться. Если «добивка» слишком длинная, она обрезается с конца.
-const concatStrings = (string,length,addString) => string.padStart(length, addString);
-
